# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sapphire is a platform for image annotation and dataset management.",
        "title": "Sapphire Server API",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/annotate/{set_id}": {
            "get": {
                "description": "Get images for annotation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "annotation"
                ],
                "summary": "Get images for annotation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dataset ID",
                        "name": "set_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of images",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}